Make DeleteChannelPermission work the same way as the Update
