+++ http/src/request/channel/mod.rs
@@ -8,6 +8,7 @@ mod create_pin;
 mod create_typing_trigger;
 mod delete_channel;
 mod delete_channel_permission;
+mod delete_channel_permission_configured;
 mod delete_pin;
 mod get_channel;
 mod get_pins;
@@ -17,7 +18,8 @@ mod update_channel_permission_configured;
 pub use self::{
     create_pin::CreatePin, create_typing_trigger::CreateTypingTrigger,
     delete_channel::DeleteChannel, delete_channel_permission::DeleteChannelPermission,
-    delete_pin::DeletePin, get_channel::GetChannel, get_pins::GetPins,
-    update_channel::UpdateChannel, update_channel_permission::UpdateChannelPermission,
+    delete_channel_permission_configured::DeleteChannelPermissionConfigured, delete_pin::DeletePin,
+    get_channel::GetChannel, get_pins::GetPins, update_channel::UpdateChannel,
+    update_channel_permission::UpdateChannelPermission,
     update_channel_permission_configured::UpdateChannelPermissionConfigured,
 };
